backend/requirements.txt
# FastAPI框架及其依赖
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 数据库
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Google Drive API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
loguru>=0.7.2

# 测试
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1 